Add files via upload
Edit 1: Contains base code and barely anything outside of video original.
Edit 2: Contains alot of the main code used for finding xp gain and gating xp. 
Final: Uses all of the previous elements to produce correct ammounts however is still a bit broken. It uses a simple equation to calculate percentage and uses that to determine win or loss.